Add files via upload
Uploading the files manually because I am not sure how this happened but the GitHub Desktop and my VS code are not communicating and it is not committing any changes I save or commit to the desktop.  The changes made are to the js page where I added an array of three Pokemon.